+++ frontend/src/main.js
@@ -2,15 +2,13 @@ import Vue from 'vue';
 import App from './App.vue';
 import router from './router';
 import axios from 'axios';
-import BootstrapVue from 'bootstrap-vue/dist/bootstrap-vue.esm';
+import BootstrapVue from 'bootstrap-vue';
 import moment from 'moment';
 import VueMomentJS from 'vue-momentjs';
 import VueSession from 'vue-session';
 
-import 'bootstrap';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap-vue/dist/bootstrap-vue.css';
-import 'expose-loader?$!expose-loader?jQuery!jquery';
 
 Vue.use(VueSession, [{ persist: true }]);
 Vue.use(BootstrapVue);